Backend_V1

+++ src/main/java/org/example/client_service/repository/AgentRepository.java
@@ -3,7 +3,7 @@ package org.example.client_service.repository;
 import org.example.client_service.models.Agent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AgentRepository extends JpaRepository<Agent, Integer> {
+public interface AgentRepository extends JpaRepository<Agent, Long> {
     Agent findByUsernameAndPassword(String username, String password);
 
 }
